services/auth/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-consul==1.1.0
pydantic==1.10.7
